Improve creation of new tree items (new feed/folder, drag&drop)
Default behavior for drag&drop and create new feed/folder is to insert new item above the selected one.
For drag&drop the shift key while dragging over a folder enable the user to insert the dragged item into the folder (as first item). when the shift key is pressed while dragging over a folder the drop marker highlights the entire folder element.
For new feed/folder the dialog provides a checkbox 'Insert inside selected folder' when selected item is a folder.