Notify the panel when to set its layout
Both views now notify the panel that their content was loaded and the panel can set its final layout. This removes the ugly 150ms delay that waited until the views loading was done.

This also enable to move the call to restoreSelectedTreeItem() from within the treeView right after the tree was created.