Modify the way filter works
Now filtering is done via CSS class 'filtered' instead of 'xxx.style.display'.

I noticed that the execution of slUtil.disableElementTree() after applying any filter is very slow. It looks like this is due to the modifications made to the HTML by the filter that addes the style attribute to the tags.