Use a hash in place of a boolean flag for CSS source indicator
prefs will now "remember" a hash of the css source inplace on a bool flag.
This hash is also an indicator to the existence of a css source. When there is no source the hash will be empty.
The hash will also be used to identify if the source has changed by comparing hashes.