Bugfix: startup tabs (pin or homepage urls) are loading in feed-preview
This happens periodically due to a synchronization issue.

The variable m_regExpRssContentTypes in background.js is initialized in a prefs.getStrictRssContentTypes() promise AFTER the webRequest.onHeadersReceived event listener is added.
If the onWebRequestHeadersReceived() listener was called while m_regExpRssContentTypes is still undefined, the test of RegExp.match() of the header's value would pass and the web request was redirected to feed-preview.

The fix was to initialize `m_regExpRssContentTypes` at the variable declaration with a default RegExp value.

Following a bug report in discourse.mozilla.org.
https://discourse.mozilla.org/t/support-sage-like-sidebar-based-rss-feed-reader/43383/32