syndication: Added SigninCredential for feeds that requires authentication (sign-in)
Currently controls when Fx will display the login dialog 'Authorization Required - Mozilla Firefox'.

Send an empty HTTP 'Authorization' header to prevent the dialog popup (fx93) in the batch processes like page discovery and feed check.

Do not send any HTTP 'Authorization' header to causes Fx to open the login dialog when user clicks on a feed in tree or open the feed-preview.

Fx96 is using the built-in password manager to fill in the above mentioned login dialog. For feeds in sidebar and for feed-preview.