discoveryView.js:  Improve; discover single feed from url if nothing was found.
If loadDiscoverFeedsList() finds nothing try to see if URL itself is the feed. This is meant mainly for XML with xslt that are loaded as regular HTML pages.

Also, when page injection fails, try to handle URLs of 'view-source'. This makes the catch() section an handler for browser build-in viewers : JSON, XML and, view-source.

Also some minor code rearranging.