Feed sorting improved by ~30%
Instead of sorting the feeder in the getFeedItems() functions, the sorting is done on the result list returned from getFeedItems(). in the syndication functions.

And since the feed-item's 'lastUpdated' property is now guaranteed to be 'Date' (#0f2ef4a) the sorting process is faster and simpler and hence the feeder sorting functions are obsolete and were removed from jsonFeed and xmlFeed.