Rewrite and Improve feed discovery.
1. common.js, background.js, content.js, pagePopup.js:
In preparation for manifest-V3 the process of the feed discovery from the address bar was rewritten.
The discovered feeds are now fetched from the background.js instead of the injected content.js. The fetched feeds are still stored in the injected content to allow repeated access from the pagePopup.

2. discoveryView.js:
Rewritten to reuse the code for discovery from the address bar to acquire the page data ('contentHandler'). Removed the no longer used 'PageDataByInjection' class.

3. discoveryView.js:
Use pageData.contentType to discover feeds in viewers.
Removed setting a default error message.

4. pagePopup.js:
Limit the repeated calls to createFeedList() using a function parameter with a default value.