Changed the way Required-Permissions are handled
Removed and rewrote many of the modifications added in commit 0ea0a36.
Actions are not blocked if the permission is not allowed. Instead the user is automatically notified with a message only once.
When fetches fail due to a NetworkError while the permission is not granted a small n in-place notification or link will be displayed.