Handling JSON comments
Decided not to include comments in the manifest JSON file. The Mozilla's submit new version process might fail. Not so critical that I should check. Moved the commented data to 'tasks.md'.

Also used the opportunity to fix 'Create-FirefoxWebExtVersion.ps1' so it can handle JSON comments. That were just removed. ¯\_(ツ)_/¯