Refactor messageView.open() in preparation for i18n
function messageView.open() can now replace '\n' with a <br> and a pair of ""' with a pair of <b> and </b> before displaying the message.
This is done to simplify the messages displayed to a more natural text and less HTML-ish.
Also refactored all text sent to messageView.open() to be less HTML-ish.

Stopped using the <u> tag for simplicity.
Still no solution for other HTML tags like <ol>, <li>, <a> or <b title=>